requirements.txt
pygnmi
pyyaml
